Qt6 Easy changes

+++ Code/Tools/AssetProcessor/native/AssetManager/assetScanFolderInfo.h
@@ -118,7 +118,7 @@ namespace AssetProcessor
         bool m_isDirectory{};
     };
 
-    inline uint qHash(const AssetFileInfo& item)
+    inline size_t qHash(const AssetFileInfo& item)
     {
         return qHash(item.m_filePath);
     }
